@ feat: HUDタブ統合・マイページのタブ化・段階別ロード画面
- ロビーとインスタンスで共通のHUDタブ(HudTabs)に統合(ホーム/ワールド/フレンド/マイページ)
- マイページを遷移なしのタブ化し /user/me ルートを廃止(/user/:id は維持)
- WorldListModal を HudOverlay にリネームし全タブ表示・周囲クリックで閉じる
- インスタンス接続を段階別ロード画面(認証/接続/参加/DL/初期化)に刷新
- ロード失敗・タイムアウト時にロビーへ戻す導線を追加

@
