+++ packages/frontend/src/components/profile/index.ts
@@ -1,0 +1,1 @@
+export { UserProfileView } from './UserProfileView';
