+++ packages/frontend/src/plugins/PluginRegistryContext.tsx
@@ -147,6 +147,12 @@ async function loadWorkerPlugin(entityType: string): Promise<WorkerPluginDefinit
 
 export type AnyPluginDefinition = WidgetDefinition | WorkerPluginDefinition;
 
+/** プラグイン（worker コード）のダウンロード進捗。total = 開始数 / completed = 完了数 */
+export interface PluginLoadingStatus {
+    completed: number;
+    total: number;
+}
+
 interface PluginRegistryContextType {
     pluginMap: Map<string, AnyPluginDefinition>;
     /** フェッチ中のプラグイン数 */
@@ -171,14 +177,15 @@ const PluginRegistryContext = createContext<PluginRegistryContextType>({
 
 export const PluginRegistryProvider: React.FC<{
     children: React.ReactNode;
-    onStatusChange?: (pendingCount: number) => void;
+    onStatusChange?: (status: PluginLoadingStatus) => void;
 }> = ({ children, onStatusChange }) => {
     const [pluginMap, setPluginMap] = useState<Map<string, AnyPluginDefinition>>(new Map());
-    const [pendingPluginCount, setPendingPluginCount] = useState(0);
+    const [loadCounts, setLoadCounts] = useState<PluginLoadingStatus>({ completed: 0, total: 0 });
+    const pendingPluginCount = loadCounts.total - loadCounts.completed;
 
     useEffect(() => {
-        onStatusChange?.(pendingPluginCount);
-    }, [pendingPluginCount, onStatusChange]);
+        onStatusChange?.(loadCounts);
+    }, [loadCounts, onStatusChange]);
     // ロード済み（またはロード中）のエンティティタイプを追跡して重複ロードを防ぐ
     const loadingRef = useRef(new Set<string>());
     // register() 呼び出し済みの plugin id を追跡（StrictMode での二重呼び出し防止）
@@ -218,7 +225,7 @@ export const PluginRegistryProvider: React.FC<{
         (entityType: string) => {
             if (loadingRef.current.has(entityType)) return;
             loadingRef.current.add(entityType);
-            setPendingPluginCount((c) => c + 1);
+            setLoadCounts((c) => ({ ...c, total: c.total + 1 }));
 
             loadWorkerPlugin(entityType)
                 .then((def) => {
@@ -241,7 +248,7 @@ export const PluginRegistryProvider: React.FC<{
                     console.error(`[PluginRegistry] Failed to load plugin: ${entityType}`, err);
                 })
                 .finally(() => {
-                    setPendingPluginCount((c) => c - 1);
+                    setLoadCounts((c) => ({ ...c, completed: c.completed + 1 }));
                 });
         },
         [addPlugin],
